Move View Model Factories to compress the packages
