--- app/src/main/java/com/melvin/ongandroid/viewmodel/ActivityViewModelFactory.kt
@@ -1,15 +1,0 @@
-package com.melvin.ongandroid.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.melvin.ongandroid.model.InicioActivitys.ActivityRemoteDataSource
-import com.melvin.ongandroid.model.InicioActivitys.ActivityRepository
-
-class ActivityViewModelFactory:ViewModelProvider.Factory {
-
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        val remoteDataSource = ActivityRemoteDataSource()
-        val repository = ActivityRepository(remoteDataSource)
-        return ActivityViewModel(repository) as T
-    }
-}
