+++ app/src/main/java/com/melvin/ongandroid/viewmodel/TestimonialsViewModel.kt
@@ -1,33 +1,34 @@
 package com.melvin.ongandroid.viewmodel
 
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.melvin.ongandroid.model.NetWorkResponse
+import androidx.lifecycle.*
+import com.melvin.ongandroid.model.OngRemoteDataSource
 import com.melvin.ongandroid.model.OngRepository
-import com.melvin.ongandroid.model.Testimonial
-import com.melvin.ongandroid.model.TestimonialsResponse
 import com.melvin.ongandroid.utils.ResultState
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 
 class TestimonialsViewModel(private val repository: OngRepository):ViewModel() {
 
-    private var _testimonialsList = MutableLiveData<ResultState<Any>>()
-    val testimonialsList : LiveData<ResultState<Any>> get() = _testimonialsList
+    private val _testimonialsResultState = MutableLiveData<ResultState<Any>>()
+    val testimonialsResultState : LiveData<ResultState<Any>> get() = _testimonialsResultState
 
     init {
         loadTestimonials()
     }
 
-
     private fun loadTestimonials(){
         viewModelScope.launch {
             repository.getTestimonialsList().collect(){ resultState ->
-                _testimonialsList.value = resultState
-
+                _testimonialsResultState.value = resultState
             }
         }
+    }
+}
+
+class TestimonialsViewModelFactory: ViewModelProvider.Factory {
+    override fun <T:ViewModel> create(modelClass: Class<T>):T{
+        val remoteDataSource = OngRemoteDataSource()
+        val repository = OngRepository(remoteDataSource)
+
+        return TestimonialsViewModel(repository) as T
     }
 }
